Create SpanRectangleOverlay specialization component
